Change purchase detail qty to decimal

+++ models/rs/purchaseDetail.model.js
@@ -10,7 +10,7 @@ module.exports = (sequelize) => {
       },
     },
     qty: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(7, 3),
       allowNull: false,
       validate: {
         min: 0,
